separate the container penjualan into several small parts
